Move cli-wallet and common rpc methods to sov-rollup-template
