+++ examples/demo-rollup/src/register_rpc.rs
@@ -8,23 +8,6 @@ use sov_rollup_interface::services::da::DaService;
 use sov_rollup_interface::zk::Zkvm;
 use sov_sequencer::get_sequencer_rpc;
 
-// register sequencer rpc methods.
-pub(crate) fn register_sequencer<Vm, Da>(
-    da_service: Da,
-    app: &mut App<Vm, Da::Spec>,
-    methods: &mut jsonrpsee::RpcModule<()>,
-) -> Result<(), anyhow::Error>
-where
-    Da: DaService,
-    Vm: Zkvm,
-{
-    let batch_builder = app.batch_builder.take().unwrap();
-    let sequencer_rpc = get_sequencer_rpc(batch_builder, da_service);
-    methods
-        .merge(sequencer_rpc)
-        .context("Failed to merge Txs RPC modules")
-}
-
 #[cfg(feature = "experimental")]
 // register ethereum methods.
 pub(crate) fn register_ethereum<C: sov_modules_api::Context, Da: DaService>(
