+++ examples/demo-rollup/src/sov-cli/main.rs
@@ -1,7 +1,13 @@
+use demo_stf::runtime::RuntimeSubcommand;
+use sov_demo_rollup::CelestiaDemoRollup;
+use sov_modules_api::cli::{FileNameArg, JsonStringArg};
+use sov_modules_rollup_template::WalletTemplate;
+
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
-    demo_stf::cli::run_wallet::<
-        <sov_celestia_adapter::CelestiaService as sov_rollup_interface::services::da::DaService>::Spec,
+    CelestiaDemoRollup::run_wallet::<
+        RuntimeSubcommand<FileNameArg, _, _>,
+        RuntimeSubcommand<JsonStringArg, _, _>,
     >()
     .await
 }
