+++ examples/demo-rollup/stf/src/lib.rs
@@ -2,8 +2,6 @@
 #![doc = include_str!("../README.md")]
 
 #[cfg(feature = "native")]
-pub mod cli;
-#[cfg(feature = "native")]
 pub mod genesis_config;
 mod hooks_impl;
 pub mod runtime;
