+++ module-system/module-implementations/integration-tests/Cargo.toml
@@ -16,6 +16,7 @@ anyhow = { workspace = true }
 borsh = { workspace = true, features = ["rc"] }
 tempfile = { workspace = true }
 serde = { workspace = true }
+jsonrpsee =  { workspace = true }
 
 sov-modules-api = { path = "../../sov-modules-api", features = ["native"] }
 sov-state = { path = "../../sov-state", features = ["native"] }
