+++ module-system/module-implementations/integration-tests/src/chain_state/helpers.rs
@@ -114,6 +114,10 @@ where
 
 impl<C: Context, Da: DaSpec> Runtime<C, Da> for TestRuntime<C, Da> {
     type GenesisConfig = GenesisConfig<C, Da>;
+
+    fn rpc_methods(_storage: <C as Spec>::Storage) -> jsonrpsee::RpcModule<()> {
+        todo!()
+    }
 }
 
 pub(crate) fn create_chain_state_genesis_config<C: Context, Da: DaSpec>(
