+++ module-system/sov-modules-rollup-template/Cargo.toml
@@ -1,6 +1,5 @@
 [package]
 name = "sov-modules-rollup-template"
-description = "Defines a generic rollup for use with the Sovereign SDK module system"
 authors = { workspace = true }
 edition = { workspace = true }
 homepage = { workspace = true }
@@ -17,10 +16,18 @@ sov-stf-runner = { path = "../../full-node/sov-stf-runner", features = ["native"
 sov-state = { path = "../sov-state", version = "0.3" }
 
 sov-modules-api = { path = "../../module-system/sov-modules-api", features = ["native"], version = "0.3" }
+sov-cli = { path = "../../module-system/sov-cli" }
+
 sov-modules-stf-template = { path = "../../module-system/sov-modules-stf-template", features = ["native"], version = "0.3" }
 sov-db = { path = "../../full-node/db/sov-db", version = "0.3" }
 
+sov-sequencer = { path = "../../full-node/sov-sequencer" }
+sov-ledger-rpc = { path = "../../full-node/sov-ledger-rpc", features = ["server"] }
+
 anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
 async-trait = { workspace = true }
 jsonrpsee = { workspace = true, features = ["http-client", "server"] }
 tokio = { workspace = true }
+borsh = { workspace = true }
