+++ sov-rollup-starter/README.md
@@ -1,5 +1,27 @@
-# Rollup Template
 
-This is still work in progress, please don't use it.
+# This is still work in progress. 
 
-This package is a convenient starting point for building a rollup using the Sovereign SDK
+This package is a convenient starting point for building a rollup using the Sovereign SDK:
+
+
+# How to run the sov-rollup-starter:
+1. Starting the node:
+If you want to run a fresh rollup remove the `rollup-starter-data` folder.
+This will compile and start the rollup node:
+
+```shell
+cargo run --bin node
+```
+
+
+2. In another shell run:
+
+```shell
+make test-create-token
+```
+
+3. Test if token creation succeeded
+
+```shell
+make test-bank-supply-of:
+```
