+++ sov-rollup-starter/stf/src/lib.rs
@@ -1,8 +1,6 @@
 //! The rollup State Transition Function.
 
 #[cfg(feature = "native")]
-pub mod cli;
-#[cfg(feature = "native")]
 pub mod genesis_config;
 mod hooks;
 pub mod runtime;
