+++ sov-rollup-starter/tests/test_helpers.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 
 use sov_modules_rollup_template::{RollupProverConfig, RollupTemplate};
 use sov_rollup_interface::mocks::{MockAddress, MockDaConfig};
-use sov_rollup_starter::rollup::StarterRollup;
+use sov_rollup_starter::StarterRollup;
 use sov_stf_runner::{RollupConfig, RpcConfig, RunnerConfig, StorageConfig};
 use stf_starter::genesis_config::GenesisPaths;
 use tokio::sync::oneshot;
